+++ js/util.js
@@ -1,0 +1,19 @@
+const getRandomInteger = (numA, numB) => {
+  const lower = Math.ceil(Math.min(numA, numB));
+  const upper = Math.floor(Math.max(numA, numB));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[Math.floor(Math.random() * elements.length)];
+
+const createIdGenerator = () => {
+  let lastGeneratedId = 0;
+
+  return function () {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+};
+
+export { getRandomInteger, getRandomArrayElement, createIdGenerator };
